requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pypodio2
